Clean up frontend structure and helper organization
