+++ static/css/exercise.css
@@ -1,0 +1,88 @@
+.activity-map {
+  height: 320px;
+  width: 100%;
+  border: 1px solid #cccccc;
+  border-radius: 10px;
+  overflow: hidden;
+}
+
+.activity-location-panel {
+  background: #f8f9fa;
+  border: 1px solid #dddddd;
+  border-radius: 10px;
+  padding: 1rem;
+}
+
+
+/* ── Legacy page styles (index / exercise log pages) ─────── */
+.top-bar {
+  background-color: #7793ae;
+  padding: 10px 30px;
+}
+
+.nav-box, .main-box, .exercise-row-box {
+  border: 1px solid #ccc;
+  background-color: white;
+}
+
+.nav-box {
+  padding: 8px 14px;
+  text-decoration: none;
+  color: black;
+  border-radius: 8px;
+  display: inline-block;
+}
+
+.main-box {
+  padding: 25px;
+}
+
+.notes-box {
+  height: 220px;
+  resize: none;
+  padding-bottom: 2rem;
+}
+
+.profile-circle {
+  width: 42px;
+  height: 42px;
+  border-radius: 50%;
+  background-color: white;
+  color: black;
+  text-decoration: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-weight: bold;
+  border: 1px solid #ccc;
+}
+
+.hover-dropdown {
+  position: relative;
+}
+
+.hover-dropdown:hover .dropdown-menu {
+  display: block;
+  margin-top: 0;
+}
+
+.dropdown-menu {
+  display: none;
+  min-width: 160px;
+}
+
+.exercise-row-box {
+  border-radius: 8px;
+  padding: 15px;
+  margin-bottom: 15px;
+}
+
+.logo {
+  display: flex;
+  align-items: center;
+}
+
+.site-logo {
+  height: 80px;
+  width: auto;
+}
