+++ static/css/history.css
@@ -1,0 +1,61 @@
+.history-map {
+  height: 360px;
+  width: 100%;
+  border: 1px solid #cccccc;
+  border-radius: 10px;
+  overflow: hidden;
+}
+
+.history-table {
+  table-layout: fixed;
+  width: 100%;
+}
+
+.history-notes {
+  width: 35%;
+}
+
+.history-notes-box {
+  max-height: 96px;
+  overflow-y: auto;
+  padding-right: 0.25rem;
+  white-space: normal;
+  overflow-wrap: anywhere;
+  word-break: break-word;
+}
+
+.history-map .leaflet-popup-content {
+  max-width: 240px;
+  overflow-wrap: anywhere;
+  word-break: break-word;
+}
+
+.activity-map .leaflet-pane,
+.activity-map .leaflet-top,
+.activity-map .leaflet-bottom,
+.history-map .leaflet-pane,
+.history-map .leaflet-top,
+.history-map .leaflet-bottom {
+  z-index: 1;
+}
+
+.empty-state {
+  border: 1px dashed #c9d3dc;
+  border-radius: 10px;
+  padding: 1.5rem;
+  background: #f8fafc;
+}
+
+.empty-state h4 {
+  margin-bottom: 0.35rem;
+}
+
+.empty-state p {
+  color: #666666;
+  margin-bottom: 1rem;
+}
+
+.empty-state-action {
+  display: inline-block;
+  text-decoration: none;
+}
