+++ static/css/ranking.css
@@ -1,0 +1,2 @@
+/* Ranking page specific styles can live here when needed. */
+
