+++ static/css/shares.css
@@ -1,0 +1,121 @@
+.shares-tab-btn {
+  background: #d0d0d0;
+  border: none;
+  border-radius: 8px;
+  padding: 0.5rem 1.25rem;
+  font-weight: 600;
+  cursor: pointer;
+  width: 100%;
+  position: relative;
+}
+
+.shares-tab-btn.active {
+  background: #555;
+  color: #fff;
+}
+
+.like-btn {
+  background: #f0f0f0;
+  border: 1.5px solid #ccc;
+  border-radius: 20px;
+  cursor: pointer;
+  font-size: 0.88rem;
+  font-weight: 600;
+  color: #888;
+  padding: 0.3rem 0.8rem;
+  align-self: center;
+  white-space: nowrap;
+  transition: background 0.15s, border-color 0.15s, color 0.15s;
+}
+
+.like-btn:hover:not(.liked) {
+  background: #ffe0ef;
+  border-color: #e91e8c;
+  color: #e91e8c;
+}
+
+.like-btn.liked {
+  background: #ffe0ef;
+  border-color: #e91e8c;
+  color: #e91e8c;
+}
+
+.post-card {
+  background: #e0e0e0;
+  border-radius: 10px;
+  padding: 0.75rem 1rem;
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.notes-wrapper {
+  position: relative;
+}
+
+.notes-limit {
+  position: absolute;
+  bottom: 10px;
+  right: 12px;
+  font-size: 0.8rem;
+  color: #666;
+  pointer-events: none;
+  background: transparent;
+}
+
+.add-record-btn {
+  width: 100%;
+}
+/* ── Shares page ──────────────────────────────────────────── */
+.shares-layout {
+  min-height: calc(100vh - 90px);
+}
+
+.shares-sidebar {
+  min-width: 120px;
+  border-right: 2px solid #000000;
+}
+
+.shares-tab-link {
+  text-decoration: none;
+}
+
+.post-meta {
+  min-width: 130px;
+}
+
+.post-date {
+  font-size: 0.92rem;
+}
+
+.post-from-to {
+  font-size: 0.9rem;
+  margin-top: 0.2rem;
+}
+
+.post-divider {
+  width: 1px;
+  background: #bbb;
+  align-self: stretch;
+}
+
+.exercise-name {
+  font-size: 0.95rem;
+  font-weight: 600;
+}
+
+.exercise-detail {
+  color: #666;
+  font-size: 0.9rem;
+}
+
+.post-total {
+  font-size: 0.9rem;
+  font-weight: 550;
+  color: #444;
+  margin-top: 0.4rem;
+}
+
+.share-liked-icon {
+  font-size: 1.3rem;
+}
