+++ templates/base.html
@@ -6,7 +6,7 @@
     <title>{% block title %}Exercise Planner{% endblock %}</title>
     <meta name="csrf-token" content="{{ csrf_token() }}">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="{{ url_for('static', filename='css/style.css') }}">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/shared.css') }}">
     {% block head %}{% endblock %}
 </head>
 <body>
