+++ templates/forgot_password.html
@@ -2,6 +2,10 @@
 
 {% block title %}Forgot Password{% endblock %}
 
+{% block head %}
+<link rel="stylesheet" href="{{ url_for('static', filename='css/auth.css') }}">
+{% endblock %}
+
 {% block navbar %}
 
 <!-- Navbar -->
