+++ templates/signup.html
@@ -2,6 +2,10 @@
 
 {% block title %}Sign Up{% endblock %}
 
+{% block head %}
+<link rel="stylesheet" href="{{ url_for('static', filename='css/auth.css') }}">
+{% endblock %}
+
 {% block content %}
 <!-- top bar -->
 <div class="auth-navbar">
