--- user_stories.md
@@ -1,36 +1,0 @@
-1.	Main page: 5/5
-As a user, I want to access the main page so that I can navigate to the app’s features easily.
-
-2.	Sign up: 5/5
-As a new user, I want to create an account so that I can save my weight loss data securely.
-
-3.	Log in: 5/5
-As a user I want to log into my account so that I can access my personal tracking information.
-
-4.	Record cardio session: 5/5
-As a user, I want to record my cardio sessions so that I can track my exercise activities over time.
-
-5.	Enter cardio detail: 5/5
-As a user, I want to enter the duration and different types of cardio exercise so that I can keep accurate workout records.
-
-6.	Log weight: 5/5
-As a user, I want to log my current weight so that I can monitor my weight loss progress.
-
-7.	View progress (dashboard): 5/5
-As a user, I want to view my exercise and weight data on the dashboard so that I can see my progress over time
-
-8.	Update profile: 3/5
-As a user, I want to edit my profile details so that I can keep my personal information up to date.
-
-9.	View achievements/ranking: 4/5
-As a user, I want to view my achievements so that I can stay motivated during my weight loss journey.
-
-10.	Notes 2/5
-As a user, I want to add notes to my cardio sessions so that I can record extra details about my workout.
-
-
-11.	Distance Tracking 3/5
-As a user, I want to log the distances that I have travelled throughout the day from my phone so that I can better track my activity levels.
-
-12.	Sharing Progress 3.5/5
-As a user, I want to see my friend’s weight loss progress so we can motivate and compete with each other.
